+++ tanks-game/defaultrandomlevel.h
@@ -4,7 +4,7 @@
 class DefaultRandomLevel : public Level {
 public:
 	std::string getName() { return "default random"; }
-	std::unordered_map<std::string, double> getWeights();
+	std::unordered_map<std::string, float> getWeights();
 	void initialize();
 
 	DefaultRandomLevel();
